Organizes constants
