+++ frontend/src/components/constants/path.tsx
@@ -1,0 +1,7 @@
+export const HOME_PATH = "/";
+export const SIGNUP_PATH = "/user/signup";
+export const SIGNIN_PATH = "/user/signin";
+export const SHOPPING_CART_PATH = "/cart";
+export const UPDATE_USER_PATH = "/user/update";
+export const ORDERS_HISTORY_PATH = "/ordershistory";
+export const SHIPPING_PATH = "/shipping";
