+++ frontend/src/components/constants/text.tsx
@@ -1,0 +1,7 @@
+export const SIGNIN_TEXT = "Signin!";
+export const SIGNUP_TEXT = "Signup!"
+export const SIGNOUT_TEXT = "Sign out!";
+export const PROFILE_TEXT = "Profile";
+export const MY_ORDERS_TEXT = "My Orders";
+export const CART_TEXT = "Cart";
+export const CURRENCY_DOLLAR = "$";
