+++ frontend/src/components/navbar/UserOptionsMobileMenu.tsx
@@ -8,7 +8,8 @@ import AccountCircle from "@mui/icons-material/AccountCircle";
 import ViewListIcon from "@mui/icons-material/ViewList";
 
 import { ShoppingCartContext } from "../../contexts/shopping-cart-context/shoppingCartContext";
-import { MY_ORDERS_TEXT, ORDERS_HISTORY_PATH, PROFILE_TEXT, SIGNOUT_TEXT, UPDATE_USER_PATH } from "../constants";
+import { ORDERS_HISTORY_PATH, UPDATE_USER_PATH } from "../constants/path";
+import { MY_ORDERS_TEXT, PROFILE_TEXT, SIGNOUT_TEXT } from "../constants/text";
 import ShoppingCart from "./ShoppingCart";
 
 const UserOptionsMobileMenu = () => {
